requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
certifi==2024.12.14
    # via
    #   -r requirements/base.txt
    #   requests
charset-normalizer==3.4.1
    # via
    #   -r requirements/base.txt
    #   requests
coverage[toml]==7.6.10
    # via pytest-cov
ddt==1.7.2
    # via -r requirements/test.in
idna==3.10
    # via
    #   -r requirements/base.txt
    #   requests
iniconfig==2.0.0
    # via pytest
packaging==24.2
    # via pytest
pluggy==1.5.0
    # via pytest
pytest==8.3.4
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==6.0.0
    # via -r requirements/test.in
pytest-django==4.9.0
    # via -r requirements/test.in
requests==2.32.3
    # via -r requirements/base.txt
urllib3==2.3.0
    # via
    #   -r requirements/base.txt
    #   requests

# The following packages are considered to be unsafe in a requirements file:
setuptools==75.6.0
    # via -r requirements/base.txt
